Add back note field tree, change search preview layout, delete old files
Rearrange docs
