+++ src/find_missing_words/gui/options.py
@@ -6,6 +6,7 @@ from .search import Search
 
 def invoke_addon_window():
     # Prevent garbage collection my assigning widget to top-level mw
+    # https://apps.ankiweb.net/docs/addons.html#qt
     mw.find_missing_words_widget = window = Search(mw)
     window.show()
